task1/app/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.1
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
